Add release.Namespace in meta and update api
* Add release.Namespace in meta

Adding release.Namespace in meta fixes issues
when rendering template with 'helm template --namespace'.
Without it, the rendered templates do not have the
namespace defined which is problematic when using
a gitops approach. This is also a requirement for
a tool like Spinnaker.

* update api versions

Use cert-manager.io/v1alpha2 and update
deprecated API for Deployment

* Bump to version 0.2.0

Signed-off-by: lcavajani <lcavajani@suse.com>